refactor(env): introduce env_model; unify reserved-env list (Phase 10)
Implements the reserved-env and interpolation-parsing parts of Phase
10. Scope narrowed vs. the plan's full wording: dependency-graph
construction and topological/cycle detection are deferred to a
follow-up because they are tightly coupled to
`manifest::EnvVarDecl` / `env_resolver::topological_sort` — extracting
them without rewriting logic would stretch the "move code before
rewriting" rule. The reserved-env unification and the interpolation
parser move cleanly as verbatim relocations.

Changes:

- New `src/env_model.rs` (172 LOC) owning:
  - `JACKIN_RUNTIME_ENV_NAME` / `JACKIN_RUNTIME_ENV_VALUE` /
    `JACKIN_DIND_HOSTNAME_ENV_NAME` public constants (moved verbatim
    from `src/manifest.rs`).
  - `RESERVED_RUNTIME_ENV_VARS` — `pub(crate)`, the **union** of the
    previous `manifest::RESERVED_RUNTIME_ENV_VARS` (5 entries) and
    `runtime::RUNTIME_OWNED_ENV_VARS` (3 entries). The manifest list
    was already a superset, so the union equals the manifest list
    unchanged.
  - `pub fn is_reserved(name: &str) -> bool` — single-line predicate
    the runtime filter now consults.
  - `extract_interpolation_refs` — `pub(crate)`, verbatim from
    `src/env_resolver.rs`.
  - Three sentinel-membership tests asserting every previously-reserved
    name is in the unified list (per the plan's mitigation).
  - Six `extract_interpolation_refs_*` tests migrated verbatim from
    `src/env_resolver.rs`.

- `src/manifest.rs`:
  - Drops the three `pub const JACKIN_*` declarations and the local
    `RESERVED_RUNTIME_ENV_VARS` private constant.
  - `pub use env_model::{JACKIN_RUNTIME_ENV_NAME,
    JACKIN_RUNTIME_ENV_VALUE, JACKIN_DIND_HOSTNAME_ENV_NAME};` so
    `crate::manifest::JACKIN_*` paths still resolve.
  - `use crate::env_resolver::extract_interpolation_refs;` →
    `use crate::env_model::extract_interpolation_refs;`.
  - Validate loop reads `env_model::RESERVED_RUNTIME_ENV_VARS`.

- `src/env_resolver.rs`:
  - Removes `extract_interpolation_refs` (moved) and its 6 tests.
  - Adds `use crate::env_model::extract_interpolation_refs;`.

- `src/runtime/mod.rs`:
  - Deletes `RUNTIME_OWNED_ENV_VARS` and its doc comment.

- `src/runtime/launch.rs`:
  - Drops `use super::RUNTIME_OWNED_ENV_VARS;`.
  - Every `crate::manifest::JACKIN_*` call site moves to
    `crate::env_model::JACKIN_*` (cleaner than routing through the
    re-export).
  - The three-line reserved-env filter
    `if key == JACKIN_RUNTIME_ENV_NAME
        || key == JACKIN_DIND_HOSTNAME_ENV_NAME
        || RUNTIME_OWNED_ENV_VARS.contains(&key.as_str())`
    collapses to `if crate::env_model::is_reserved(key)`.
    Behavior is identical: the union list covers all three previous
    branches.

Behavior:

- No CLI, flag, error-message, or schema change.
- Reserved-env membership identical to previous manifest list (5
  entries). Runtime filter now applies the same list the manifest
  validates against; previously the runtime filter covered the
  JACKIN_* names via inline checks and the Docker vars via the
  now-deleted `RUNTIME_OWNED_ENV_VARS`. Effective coverage unchanged.

Test Preservation Policy: all 418 baseline tests preserved; 3 new
sentinel tests added per the plan's explicit mitigation requirement.
Tests that moved (6 `extract_interpolation_refs_*`) kept their
assertions verbatim.

Test count: 418 → 421 (+3 sentinel tests; 0 deletions, 0 weakenings).

Visibility decisions:
- `RESERVED_RUNTIME_ENV_VARS`: `pub(crate)` (no external consumer).
- `extract_interpolation_refs`: kept at previous `pub(crate)`
  scope (not widened to `pub`).
- `is_reserved`: `pub` so runtime/launch and manifest can both use
  it cleanly.
- Three `JACKIN_*` constants: `pub` (matches previous manifest
  visibility) and re-exported from manifest for path compatibility.

Verification per COMMITS.md:
- cargo fmt -- --check — pass
- cargo clippy — zero warnings
- cargo nextest run — 421 tests run: 421 passed, 0 skipped

Deferred for follow-up:
- Unify `manifest::detect_env_cycles` and `env_resolver::topological_sort`
  into env_model. Both operate on `EnvVarDecl`-shaped data and need a
  small shape adapter, which is "rewrite" not "move." Separate PR.
